+++ src/main/java/pl/tw/ksiegowosc/dto/AllegroClientCredentials.java
@@ -2,6 +2,9 @@ package pl.tw.ksiegowosc.dto;
 
 public record AllegroClientCredentials(
         String clientId,
-        String clientSecret
+        String clientSecret,
+        String apiBaseUrl,
+        String authUrl,
+        String userAgent
 ) {
 }